refactor(engine-host): move canonical-cwd, context-overflow, and mcp-shutdown below the seam
Move the last of the prompt/auth/context helper set out of main.rs:
- canonical_session_cwd + context_overflow_user_message -> engine_host::session
  (they operate on runtime::Session / the session cwd).
- shutdown_mcp_state_best_effort -> engine_host::mcp (beside RuntimeMcpState).

canonical_session_cwd now returns Result<_, String> instead of
engine_acp::AcpError (engine-host stays independent of engine-acp); the two ACP
delegate call sites wrap it via .map_err(AcpError::invalid_params), and the
String-returning SessionEngine::build path drops its now-redundant
.map_err(|e| e.to_string()). context_overflow_user_message already returned a
String (its callers do the AcpError wrapping), so no call-site change there.
No new engine-host deps (runtime's CompactionConfig is already available).

Fifth staged step (part c of c) of the engine-host extraction (#120). With this,
CliHookProgressReporter (a direct-stderr renderer sink) is the only prompt/auth/
context helper left in the CLI, by design.
